requirements.txt
Flask
tensorflow
tensorflow-hub
Pillow
